Move feathers types inline
